requirements.txt
python-telegram-bot==20.0
requests==2.28.1
python-dotenv==0.19.0
urllib3==1.26.0